+++ src/shared/webpack-settings.js
@@ -6,7 +6,7 @@ export default function getWebpackSettings(server: boolean) {
 
   return {
     output: {
-      path: `${frameworkConfig.directories.build}/webpack-${server ? 'server' : 'client'}`,
+      path: `${frameworkConfig.directories.build}/${server ? 'server' : 'client'}`,
       publicPath: '/',
     },
   };